# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Irish Payroll Management & Compliance System",
  "description": "Portfolio prototype. Statutory values are versioned rules with official sources; unverified rules are blocked from calculation. Not certified payroll software.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/rules/tax": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Tax Rules",
    "operationId": "tax_rules_rules_tax_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/rules/usc": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Usc Rules",
    "operationId": "usc_rules_rules_usc_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/rules/prsi": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Prsi Rules",
    "operationId": "prsi_rules_rules_prsi_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/rules": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "All Rules",
    "operationId": "all_rules_rules_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/rules/{rule_id}/status": {
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "Rule Status",
    "operationId": "rule_status_rules__rule_id__status_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rule Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Status"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReasonIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/analytics/payroll": {
   "get": {
    "tags": [
     "analytics"
    ],
    "summary": "Payroll Analytics",
    "operationId": "payroll_analytics_analytics_payroll_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "company_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Company Id"
      }
     },
     {
      "name": "tax_year",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 2026,
       "title": "Tax Year"
      }
     },
     {
      "name": "period",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Period"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/revenue/submission/prepare": {
   "post": {
    "tags": [
     "revenue"
    ],
    "summary": "Revenue Prepare",
    "operationId": "revenue_prepare_revenue_submission_prepare_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/revenue/submission/{submission_id}/export": {
   "get": {
    "tags": [
     "revenue"
    ],
    "summary": "Revenue Export",
    "operationId": "revenue_export_revenue_submission__submission_id__export_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "submission_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Submission Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/audit": {
   "get": {
    "tags": [
     "audit"
    ],
    "summary": "Audit Log",
    "operationId": "audit_log_audit_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "action",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Action"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ReasonIn": {
    "properties": {
     "reason": {
      "type": "string",
      "minLength": 5,
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "reason"
    ],
    "title": "ReasonIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/auth/token"
     }
    }
   }
  }
 }
}